Add files via upload
Reworked the script to now include the AI having sirens and lights on emergency vehicles, 
Also the AI Ped gets back into the vehicle and drives away before despawning.
Another enhancement is a teleport, because sometimes the medic cannot get to you if you are in a tricky spot or in a vehicle, so after a 1 minute period of not successfully being able to get to you, you are brought to him to be revived. This last part is experimental and might need a few tweaks, So report any issues please.
Enjoy!